Actuator disk compiles
