+++ amr-wind/wind_energy/actuator/turbine/turbine_types.H
@@ -35,6 +35,7 @@ struct ComponentView
     RealSlice chord;
 };
 
+using DeviceVecComponent = amrex::Gpu::DeviceVector<ComponentView>;
 struct TurbineBaseData
 {
     //! Number of blades
